+++ src/pages/Mission.vue
@@ -168,14 +168,14 @@
           <q-td key="paramName" :props="props"
                 style="width:5px;font-family:monospace;vertical-align:top"
           >
-            <router-link
-              style="cursor:pointer"
-              tag="div"
-              :class="props.row.required ? 'text-bold' : 'text-black'"
+            <q-btn
+              no-caps dense
+              style="min-width:3em"
+              :class="`text-primary ${props.row.required ? 'text-bold' : ''}`"
               :to="`/${encodeURIComponent(mission.executorId)}/missiontpls/${encodeURIComponent(mission.missionTplId)}/params/${encodeURIComponent(props.row.paramName)}`"
-            >{{ props.row.paramName }}
-            </router-link>
-
+            >
+              {{ props.row.paramName }}
+            </q-btn>
             <div class="text-grey-7 q-mt-sm" style="font-size:0.8em">
               ({{ props.row.type }})
             </div>