+++ src/pages/MissionTpl.vue
@@ -105,17 +105,17 @@
           <q-td key="paramName" :props="props"
                 style="width:5px;font-family:monospace;vertical-align:top"
           >
-            <router-link
-              style="cursor:pointer"
-              tag="div"
-              :class="props.row.required ? 'text-bold' : 'text-black'"
+            <q-btn
+              no-caps dense
+              style="min-width:3em"
+              :class="`text-primary ${props.row.required ? 'text-bold' : ''}`"
               :to="`/${encodeURIComponent(params.executorId)}/missiontpls/${encodeURIComponent(params.missionTplId)}/params/${encodeURIComponent(props.row.paramName)}`"
             >
               {{ props.row.paramName }}
-              <div class="text-grey-7 q-mt-sm" style="font-size:0.8em">
-                ({{ props.row.type }})
-              </div>
-            </router-link>
+            </q-btn>
+            <div class="text-grey-7 q-mt-sm" style="font-size:0.8em">
+              ({{ props.row.type }})
+            </div>
           </q-td>
 
           <q-td key="defaultValue" :props="props"