+++ ChangeLog.md
@@ -5,8 +5,8 @@
     - geo type: also allow editing of the text?
 - TODO add Duration type  
 - TODO define how to represent tethysdash parameter types  
-- TODO delete a parameter     
     
+- add option to delete a parameter     
 - handle edit dispatch through parameter-value
 - check for args with error before validating or running the mission
   against executor system